refactor(runtime): extract security types into advanced_options module
Move SecurityOptions, ResourceLimits, SecurityOptionsBuilder, and
AdvancedBoxOptions from options.rs into a dedicated advanced_options.rs
module. Rename AdvancedOptions to AdvancedBoxOptions to reflect its
association with BoxOptions.

Eliminate re-export patterns — all imports now reference source modules
directly. Delete jailer/config.rs (was a pure re-export file).

Also adds SECCOMP_FILTER_FLAG_TSYNC support for thread-synchronized
seccomp filters on Linux.
