+++ boxlite/src/runtime/advanced_options.rs
@@ -1,0 +1,521 @@
+//! Advanced options for expert users.
+//!
+//! This module contains [`AdvancedBoxOptions`], [`SecurityOptions`], [`ResourceLimits`],
+//! and [`SecurityOptionsBuilder`] — configuration that entry-level users can safely
+//! ignore. Defaults are secure on all supported platforms.
+
+use serde::{Deserialize, Serialize};
+use std::path::PathBuf;
+
+// ============================================================================
+// Security Options
+// ============================================================================
+
+/// Security isolation options for a box.
+///
+/// These options control how the boxlite-shim process is isolated from the host.
+/// Different presets are available for different security requirements.
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct SecurityOptions {
+    /// Enable jailer isolation.
+    ///
+    /// When true, applies platform-specific security isolation:
+    /// - Linux: seccomp, namespaces, chroot, privilege drop
+    /// - macOS: sandbox-exec profile
+    ///
+    /// Default: true on Linux/macOS (use `SecurityOptions::development()` to disable)
+    #[serde(default = "default_jailer_enabled")]
+    pub jailer_enabled: bool,
+
+    /// Enable seccomp syscall filtering (Linux only).
+    ///
+    /// When true, applies a whitelist of allowed syscalls.
+    /// Default: true on Linux (use `SecurityOptions::development()` to disable)
+    #[serde(default = "default_seccomp_enabled")]
+    pub seccomp_enabled: bool,
+
+    /// UID to drop to after setup (Linux only).
+    ///
+    /// - None: Auto-allocate an unprivileged UID
+    /// - Some(0): Don't drop privileges (not recommended)
+    /// - Some(uid): Drop to specific UID
+    #[serde(default)]
+    pub uid: Option<u32>,
+
+    /// GID to drop to after setup (Linux only).
+    ///
+    /// - None: Auto-allocate an unprivileged GID
+    /// - Some(0): Don't drop privileges (not recommended)
+    /// - Some(gid): Drop to specific GID
+    #[serde(default)]
+    pub gid: Option<u32>,
+
+    /// Create new PID namespace (Linux only).
+    ///
+    /// When true, the shim becomes PID 1 in a new namespace.
+    /// Default: false
+    #[serde(default)]
+    pub new_pid_ns: bool,
+
+    /// Create new network namespace (Linux only).
+    ///
+    /// When true, creates isolated network namespace.
+    /// Note: gvproxy handles networking, so this may not be needed.
+    /// Default: false
+    #[serde(default)]
+    pub new_net_ns: bool,
+
+    /// Base directory for chroot jails (Linux only).
+    ///
+    /// Default: /srv/boxlite
+    #[serde(default = "default_chroot_base")]
+    pub chroot_base: PathBuf,
+
+    /// Enable chroot isolation (Linux only).
+    ///
+    /// When true, uses pivot_root to isolate filesystem.
+    /// Default: true on Linux
+    #[serde(default = "default_chroot_enabled")]
+    pub chroot_enabled: bool,
+
+    /// Close inherited file descriptors.
+    ///
+    /// When true, closes all FDs except stdin/stdout/stderr before VM start.
+    /// Default: true
+    #[serde(default = "default_close_fds")]
+    pub close_fds: bool,
+
+    /// Sanitize environment variables.
+    ///
+    /// When true, clears all environment variables except those in allowlist.
+    /// Default: true
+    #[serde(default = "default_sanitize_env")]
+    pub sanitize_env: bool,
+
+    /// Environment variables to preserve when sanitizing.
+    ///
+    /// Default: ["RUST_LOG", "PATH", "HOME", "USER", "LANG"]
+    #[serde(default = "default_env_allowlist")]
+    pub env_allowlist: Vec<String>,
+
+    /// Resource limits to apply.
+    #[serde(default)]
+    pub resource_limits: ResourceLimits,
+
+    /// Custom sandbox profile path (macOS only).
+    ///
+    /// If None, uses the built-in modular sandbox profile.
+    #[serde(default)]
+    pub sandbox_profile: Option<PathBuf>,
+
+    /// Enable network access in sandbox (macOS only).
+    ///
+    /// When true, adds network policy to the sandbox.
+    /// Default: true (needed for gvproxy VM networking)
+    #[serde(default = "default_network_enabled")]
+    pub network_enabled: bool,
+}
+
+/// Resource limits for the jailed process.
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
+pub struct ResourceLimits {
+    /// Maximum number of open file descriptors (RLIMIT_NOFILE).
+    #[serde(default)]
+    pub max_open_files: Option<u64>,
+
+    /// Maximum file size in bytes (RLIMIT_FSIZE).
+    #[serde(default)]
+    pub max_file_size: Option<u64>,
+
+    /// Maximum number of processes (RLIMIT_NPROC).
+    #[serde(default)]
+    pub max_processes: Option<u64>,
+
+    /// Maximum virtual memory in bytes (RLIMIT_AS).
+    #[serde(default)]
+    pub max_memory: Option<u64>,
+
+    /// Maximum CPU time in seconds (RLIMIT_CPU).
+    #[serde(default)]
+    pub max_cpu_time: Option<u64>,
+}
+
+// Default value functions for SecurityOptions
+
+fn default_jailer_enabled() -> bool {
+    cfg!(any(target_os = "linux", target_os = "macos"))
+}
+
+fn default_seccomp_enabled() -> bool {
+    cfg!(target_os = "linux")
+}
+
+fn default_chroot_base() -> PathBuf {
+    PathBuf::from("/srv/boxlite")
+}
+
+fn default_chroot_enabled() -> bool {
+    cfg!(target_os = "linux")
+}
+
+fn default_close_fds() -> bool {
+    true
+}
+
+fn default_sanitize_env() -> bool {
+    true
+}
+
+fn default_env_allowlist() -> Vec<String> {
+    vec![
+        "RUST_LOG".to_string(),
+        "PATH".to_string(),
+        "HOME".to_string(),
+        "USER".to_string(),
+        "LANG".to_string(),
+        "TERM".to_string(),
+    ]
+}
+
+fn default_network_enabled() -> bool {
+    true
+}
+
+impl Default for SecurityOptions {
+    fn default() -> Self {
+        Self {
+            jailer_enabled: default_jailer_enabled(),
+            seccomp_enabled: default_seccomp_enabled(),
+            uid: None,
+            gid: None,
+            new_pid_ns: false,
+            new_net_ns: false,
+            chroot_base: default_chroot_base(),
+            chroot_enabled: default_chroot_enabled(),
+            close_fds: default_close_fds(),
+            sanitize_env: default_sanitize_env(),
+            env_allowlist: default_env_allowlist(),
+            resource_limits: ResourceLimits::default(),
+            sandbox_profile: None,
+            network_enabled: default_network_enabled(),
+        }
+    }
+}
+
+impl SecurityOptions {
+    /// Development mode: minimal isolation for debugging.
+    ///
+    /// Use this when debugging issues where isolation interferes.
+    pub fn development() -> Self {
+        Self {
+            jailer_enabled: false,
+            seccomp_enabled: false,
+            chroot_enabled: false,
+            close_fds: false,
+            sanitize_env: false,
+            ..Default::default()
+        }
+    }
+
+    /// Standard mode: recommended for most use cases.
+    ///
+    /// Equivalent to `SecurityOptions::default()`. Provides good security
+    /// without being overly restrictive. Enables jailer on Linux/macOS,
+    /// seccomp on Linux.
+    pub fn standard() -> Self {
+        Self::default()
+    }
+
+    /// Maximum mode: all isolation features enabled.
+    ///
+    /// Use this for untrusted workloads (AI sandbox, multi-tenant).
+    pub fn maximum() -> Self {
+        Self {
+            jailer_enabled: true,
+            seccomp_enabled: cfg!(target_os = "linux"),
+            uid: Some(65534), // nobody
+            gid: Some(65534), // nogroup
+            new_pid_ns: cfg!(target_os = "linux"),
+            new_net_ns: false, // gvproxy needs network
+            chroot_enabled: cfg!(target_os = "linux"),
+            close_fds: true,
+            sanitize_env: true,
+            env_allowlist: vec!["RUST_LOG".to_string()],
+            resource_limits: ResourceLimits {
+                max_open_files: Some(1024),
+                max_file_size: Some(1024 * 1024 * 1024), // 1GB
+                max_processes: Some(100),
+                max_memory: None,   // Let VM config handle this
+                max_cpu_time: None, // Let VM config handle this
+            },
+            ..Default::default()
+        }
+    }
+
+    /// Check if current platform supports full jailer features.
+    pub fn is_full_isolation_available() -> bool {
+        cfg!(target_os = "linux")
+    }
+
+    /// Create a builder for customizing security options.
+    ///
+    /// Starts with default settings (jailer enabled on Linux/macOS).
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use boxlite::runtime::advanced_options::SecurityOptions;
+    ///
+    /// let security = SecurityOptions::builder()
+    ///     .max_open_files(1024)
+    ///     .build();
+    /// ```
+    pub fn builder() -> SecurityOptionsBuilder {
+        SecurityOptionsBuilder::new()
+    }
+}
+
+// ============================================================================
+// Security Options Builder (C-BUILDER: Non-consuming builder pattern)
+// ============================================================================
+
+/// Builder for customizing [`SecurityOptions`].
+///
+/// Provides a fluent API for configuring security isolation options.
+/// Uses non-consuming methods per Rust API guidelines (C-BUILDER).
+///
+/// # Example
+///
+/// ```
+/// use boxlite::runtime::advanced_options::SecurityOptionsBuilder;
+///
+/// let security = SecurityOptionsBuilder::standard()
+///     .max_open_files(2048)
+///     .max_file_size_bytes(1024 * 1024 * 512) // 512 MiB
+///     .build();
+/// ```
+#[derive(Debug, Clone)]
+pub struct SecurityOptionsBuilder {
+    inner: SecurityOptions,
+}
+
+impl Default for SecurityOptionsBuilder {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl SecurityOptionsBuilder {
+    /// Create a builder starting from default options.
+    pub fn new() -> Self {
+        Self {
+            inner: SecurityOptions::default(),
+        }
+    }
+
+    /// Create a builder starting from development settings.
+    ///
+    /// Minimal isolation for debugging.
+    pub fn development() -> Self {
+        Self {
+            inner: SecurityOptions::development(),
+        }
+    }
+
+    /// Create a builder starting from standard settings.
+    ///
+    /// Recommended for most use cases.
+    pub fn standard() -> Self {
+        Self {
+            inner: SecurityOptions::standard(),
+        }
+    }
+
+    /// Create a builder starting from maximum security settings.
+    ///
+    /// All isolation features enabled.
+    pub fn maximum() -> Self {
+        Self {
+            inner: SecurityOptions::maximum(),
+        }
+    }
+
+    // ─────────────────────────────────────────────────────────────────────
+    // Core isolation settings
+    // ─────────────────────────────────────────────────────────────────────
+
+    /// Enable or disable jailer isolation.
+    pub fn jailer_enabled(&mut self, enabled: bool) -> &mut Self {
+        self.inner.jailer_enabled = enabled;
+        self
+    }
+
+    /// Enable or disable seccomp syscall filtering (Linux only).
+    pub fn seccomp_enabled(&mut self, enabled: bool) -> &mut Self {
+        self.inner.seccomp_enabled = enabled;
+        self
+    }
+
+    /// Set UID to drop to after setup (Linux only).
+    pub fn uid(&mut self, uid: u32) -> &mut Self {
+        self.inner.uid = Some(uid);
+        self
+    }
+
+    /// Set GID to drop to after setup (Linux only).
+    pub fn gid(&mut self, gid: u32) -> &mut Self {
+        self.inner.gid = Some(gid);
+        self
+    }
+
+    /// Enable or disable new PID namespace (Linux only).
+    pub fn new_pid_ns(&mut self, enabled: bool) -> &mut Self {
+        self.inner.new_pid_ns = enabled;
+        self
+    }
+
+    /// Enable or disable new network namespace (Linux only).
+    pub fn new_net_ns(&mut self, enabled: bool) -> &mut Self {
+        self.inner.new_net_ns = enabled;
+        self
+    }
+
+    // ─────────────────────────────────────────────────────────────────────
+    // Filesystem isolation
+    // ─────────────────────────────────────────────────────────────────────
+
+    /// Set base directory for chroot jails (Linux only).
+    pub fn chroot_base(&mut self, path: impl Into<PathBuf>) -> &mut Self {
+        self.inner.chroot_base = path.into();
+        self
+    }
+
+    /// Enable or disable chroot isolation (Linux only).
+    pub fn chroot_enabled(&mut self, enabled: bool) -> &mut Self {
+        self.inner.chroot_enabled = enabled;
+        self
+    }
+
+    /// Enable or disable closing inherited file descriptors.
+    pub fn close_fds(&mut self, enabled: bool) -> &mut Self {
+        self.inner.close_fds = enabled;
+        self
+    }
+
+    // ─────────────────────────────────────────────────────────────────────
+    // Environment settings
+    // ─────────────────────────────────────────────────────────────────────
+
+    /// Enable or disable environment variable sanitization.
+    pub fn sanitize_env(&mut self, enabled: bool) -> &mut Self {
+        self.inner.sanitize_env = enabled;
+        self
+    }
+
+    /// Set environment variables to preserve when sanitizing.
+    pub fn env_allowlist(&mut self, vars: Vec<String>) -> &mut Self {
+        self.inner.env_allowlist = vars;
+        self
+    }
+
+    /// Add an environment variable to the allowlist.
+    pub fn allow_env(&mut self, var: impl Into<String>) -> &mut Self {
+        self.inner.env_allowlist.push(var.into());
+        self
+    }
+
+    // ─────────────────────────────────────────────────────────────────────
+    // Resource limits (type-safe setters)
+    // ─────────────────────────────────────────────────────────────────────
+
+    /// Set all resource limits at once.
+    pub fn resource_limits(&mut self, limits: ResourceLimits) -> &mut Self {
+        self.inner.resource_limits = limits;
+        self
+    }
+
+    /// Set maximum number of open file descriptors.
+    pub fn max_open_files(&mut self, limit: u64) -> &mut Self {
+        self.inner.resource_limits.max_open_files = Some(limit);
+        self
+    }
+
+    /// Set maximum file size in bytes.
+    pub fn max_file_size_bytes(&mut self, bytes: u64) -> &mut Self {
+        self.inner.resource_limits.max_file_size = Some(bytes);
+        self
+    }
+
+    /// Set maximum number of processes.
+    pub fn max_processes(&mut self, limit: u64) -> &mut Self {
+        self.inner.resource_limits.max_processes = Some(limit);
+        self
+    }
+
+    /// Set maximum virtual memory in bytes.
+    pub fn max_memory_bytes(&mut self, bytes: u64) -> &mut Self {
+        self.inner.resource_limits.max_memory = Some(bytes);
+        self
+    }
+
+    /// Set maximum CPU time in seconds.
+    pub fn max_cpu_time_seconds(&mut self, seconds: u64) -> &mut Self {
+        self.inner.resource_limits.max_cpu_time = Some(seconds);
+        self
+    }
+
+    // ─────────────────────────────────────────────────────────────────────
+    // macOS-specific settings
+    // ─────────────────────────────────────────────────────────────────────
+
+    /// Set custom sandbox profile path (macOS only).
+    pub fn sandbox_profile(&mut self, path: impl Into<PathBuf>) -> &mut Self {
+        self.inner.sandbox_profile = Some(path.into());
+        self
+    }
+
+    /// Enable or disable network access in sandbox (macOS only).
+    pub fn network_enabled(&mut self, enabled: bool) -> &mut Self {
+        self.inner.network_enabled = enabled;
+        self
+    }
+
+    // ─────────────────────────────────────────────────────────────────────
+    // Build
+    // ─────────────────────────────────────────────────────────────────────
+
+    /// Build the configured [`SecurityOptions`].
+    pub fn build(&self) -> SecurityOptions {
+        self.inner.clone()
+    }
+}
+
+// ============================================================================
+// Advanced Options
+// ============================================================================
+
+/// Advanced options for expert users.
+///
+/// Entry-level users can ignore this — defaults are secure and reasonable.
+/// Only modify these if you understand the security implications.
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
+pub struct AdvancedBoxOptions {
+    /// Security isolation options (jailer, seccomp, namespaces, resource limits).
+    ///
+    /// Defaults are secure on all supported platforms. Use presets:
+    /// - `SecurityOptions::default()` / `standard()` — recommended for production
+    /// - `SecurityOptions::development()` — minimal isolation for debugging
+    /// - `SecurityOptions::maximum()` — maximum isolation for untrusted workloads
+    #[serde(default)]
+    pub security: SecurityOptions,
+
+    /// Enable bind mount isolation for the shared mounts directory.
+    ///
+    /// When true, creates a read-only bind mount from `mounts/` to `shared/`,
+    /// preventing the guest from modifying host-prepared files.
+    ///
+    /// Requires CAP_SYS_ADMIN (privileged) or FUSE (rootless) on Linux.
+    /// Defaults to false.
+    #[serde(default)]
+    pub isolate_mounts: bool,
+}
