+++ boxlite/src/vmm/controller/shim.rs
@@ -270,7 +270,7 @@ impl VmmController for ShimController {
         let serializable_config = InstanceSpec {
             // Box identification and security (from ShimController)
             box_id: self.box_id.to_string(),
-            security: self.options.security.clone(),
+            security: self.options.advanced.security.clone(),
             // VM configuration
             cpus: config.cpus,
             memory_mib: config.memory_mib,
