+++ sdks/python/src/lib.rs
@@ -12,7 +12,9 @@ use crate::box_handle::PyBox;
 use crate::exec::{PyExecStderr, PyExecStdin, PyExecStdout, PyExecution};
 use crate::info::{PyBoxInfo, PyBoxStateInfo};
 use crate::metrics::{PyBoxMetrics, PyRuntimeMetrics};
-use crate::options::{PyBoxOptions, PyCopyOptions, PyOptions, PySecurityOptions};
+use crate::options::{
+    PyAdvancedBoxOptions, PyBoxOptions, PyCopyOptions, PyOptions, PySecurityOptions,
+};
 use crate::runtime::PyBoxlite;
 use pyo3::prelude::*;
 
@@ -21,6 +23,7 @@ fn boxlite_python(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<PyOptions>()?;
     m.add_class::<PyBoxOptions>()?;
     m.add_class::<PySecurityOptions>()?;
+    m.add_class::<PyAdvancedBoxOptions>()?;
     m.add_class::<PyBoxlite>()?;
     m.add_class::<PyBox>()?;
     m.add_class::<PyExecution>()?;
